+++ mex/backend/graph/connector.py
@@ -13,7 +13,11 @@ from neo4j import (
 from neo4j.exceptions import Neo4jError
 from pydantic import Field
 
-from mex.backend.fields import SEARCHABLE_CLASSES, SEARCHABLE_FIELDS
+from mex.backend.fields import (
+    ALL_REFERENCE_FIELD_NAMES,
+    SEARCHABLE_CLASSES,
+    SEARCHABLE_FIELDS,
+)
 from mex.backend.graph.exceptions import InconsistentGraphError, IngestionError
 from mex.backend.graph.models import IngestData, Result
 from mex.backend.graph.query import Query, QueryBuilder
@@ -117,24 +121,22 @@ class GraphConnector(BaseConnector):
             raise MExError(msg) from None
         return result
 
-    def _seed_constraints(self) -> list[Result]:
-        """Ensure uniqueness constraints are enabled for all entity types."""
+    def _seed_constraints(self) -> None:
+        """Ensure property constraints are created for all entity types."""
         query_builder = QueryBuilder.get()
         with self.driver.session(default_access_mode=WRITE_ACCESS) as session:
-            results = [
+            for label in EXTRACTED_MODEL_CLASSES_BY_NAME | MERGED_MODEL_CLASSES_BY_NAME:
                 self.commit(
-                    query_builder.create_identifier_uniqueness_constraint(
-                        node_label=class_name
-                    ),
+                    query_builder.create_identifier_constraint(node_label=label),
                     session=session,
                 )
-                for class_name in sorted(
-                    set(EXTRACTED_MODEL_CLASSES_BY_NAME)
-                    | set(MERGED_MODEL_CLASSES_BY_NAME)
+            logger.info("seeded identifier constraints")
+            for label in EXTRACTED_MODEL_CLASSES_BY_NAME:
+                self.commit(
+                    query_builder.create_provenance_constraint(node_label=label),
+                    session=session,
                 )
-            ]
-        logger.info("seeded identifier uniqueness constraints")
-        return results
+            logger.info("seeded provenance constraints")
 
     def _seed_indices(self) -> Result:
         """Ensure there is a full text search index for all searchable fields."""
@@ -194,7 +196,8 @@ class GraphConnector(BaseConnector):
         identifier: str | None,
         stable_target_id: str | None,
         entity_type: Sequence[str],
-        had_primary_source: Sequence[str] | None,
+        referenced_identifiers: Sequence[str] | None,
+        reference_field: str | None,
         skip: int,
         limit: int,
     ) -> Result:
@@ -205,7 +208,8 @@ class GraphConnector(BaseConnector):
             identifier: Optional identifier filter
             stable_target_id: Optional stable target ID filter
             entity_type: List of allowed entity types
-            had_primary_source: Optional merged primary source identifier filter
+            referenced_identifiers: Optional merged item identifiers filter
+            reference_field: Optional field name to filter for
             skip: How many items to skip for pagination
             limit: How many items to return at most
 
@@ -217,8 +221,8 @@ class GraphConnector(BaseConnector):
             filter_by_query_string=bool(query_string),
             filter_by_identifier=bool(identifier),
             filter_by_stable_target_id=bool(stable_target_id),
-            filter_by_reference_to_merged_item=bool(had_primary_source),
-            reference_field_name="hadPrimarySource",
+            filter_by_referenced_identifiers=bool(referenced_identifiers),
+            reference_field=reference_field,
         )
         result = self.commit(
             query,
@@ -226,7 +230,7 @@ class GraphConnector(BaseConnector):
             identifier=identifier,
             stable_target_id=stable_target_id,
             labels=entity_type,
-            referenced_identifiers=had_primary_source,
+            referenced_identifiers=referenced_identifiers,
             skip=skip,
             limit=limit,
         )
@@ -241,7 +245,8 @@ class GraphConnector(BaseConnector):
         identifier: str | None,
         stable_target_id: str | None,
         entity_type: Sequence[str] | None,
-        had_primary_source: Sequence[str] | None,
+        referenced_identifiers: Sequence[str] | None,
+        reference_field: str | None,
         skip: int,
         limit: int,
     ) -> Result:
@@ -252,7 +257,8 @@ class GraphConnector(BaseConnector):
             identifier: Optional identifier filter
             stable_target_id: Optional stable target ID filter
             entity_type: Optional entity type filter
-            had_primary_source: Optional merged primary source identifier filter
+            referenced_identifiers: Optional merged item identifiers filter
+            reference_field: Optional field name to filter for
             skip: How many items to skip for pagination
             limit: How many items to return at most
 
@@ -264,7 +270,8 @@ class GraphConnector(BaseConnector):
             identifier=identifier,
             stable_target_id=stable_target_id,
             entity_type=entity_type or list(EXTRACTED_MODEL_CLASSES_BY_NAME),
-            had_primary_source=had_primary_source,
+            referenced_identifiers=referenced_identifiers,
+            reference_field=reference_field,
             skip=skip,
             limit=limit,
         )
@@ -275,7 +282,8 @@ class GraphConnector(BaseConnector):
         identifier: str | None,
         stable_target_id: str | None,
         entity_type: Sequence[str] | None,
-        had_primary_source: Sequence[str] | None,
+        referenced_identifiers: Sequence[str] | None,
+        reference_field: str | None,
         skip: int,
         limit: int,
     ) -> Result:
@@ -286,7 +294,8 @@ class GraphConnector(BaseConnector):
             identifier: Optional identifier filter
             stable_target_id: Optional stable target ID filter
             entity_type: Optional entity type filter
-            had_primary_source: Optional merged primary source identifier filter
+            referenced_identifiers: Optional merged item identifiers filter
+            reference_field: Optional field name to filter for
             skip: How many items to skip for pagination
             limit: How many items to return at most
 
@@ -298,7 +307,8 @@ class GraphConnector(BaseConnector):
             identifier=identifier,
             stable_target_id=stable_target_id,
             entity_type=entity_type or list(RULE_MODEL_CLASSES_BY_NAME),
-            had_primary_source=had_primary_source,
+            referenced_identifiers=referenced_identifiers,
+            reference_field=reference_field,
             skip=skip,
             limit=limit,
         )
@@ -308,7 +318,8 @@ class GraphConnector(BaseConnector):
         query_string: str | None,
         identifier: str | None,
         entity_type: Sequence[str] | None,
-        had_primary_source: Sequence[str] | None,
+        referenced_identifiers: Sequence[str] | None,
+        reference_field: str | None,
         skip: int,
         limit: int,
     ) -> Result:
@@ -318,26 +329,30 @@ class GraphConnector(BaseConnector):
             query_string: Optional full text search query term
             identifier: Optional merged item identifier filter
             entity_type: Optional merged entity type filter
-            had_primary_source: Optional merged primary source identifier filter
+            referenced_identifiers: Optional merged item identifiers filter
+            reference_field: Optional field name to filter for
             skip: How many items to skip for pagination
             limit: How many items to return at most
 
         Returns:
             Graph result instance
         """
+        if reference_field and reference_field not in ALL_REFERENCE_FIELD_NAMES:
+            msg = "Invalid field name."
+            raise ValueError(msg)
         query_builder = QueryBuilder.get()
         query = query_builder.fetch_merged_items(
             filter_by_query_string=bool(query_string),
             filter_by_identifier=bool(identifier),
-            filter_by_reference_to_merged_item=bool(had_primary_source),
-            reference_field_name="hadPrimarySource",
+            filter_by_referenced_identifiers=bool(referenced_identifiers),
+            reference_field=reference_field,
         )
         result = self.commit(
             query,
             query_string=query_string,
             identifier=identifier,
             labels=entity_type or list(MERGED_MODEL_CLASSES_BY_NAME),
-            referenced_identifiers=had_primary_source,
+            referenced_identifiers=referenced_identifiers,
             skip=skip,
             limit=limit,
         )