+++ mex/backend/settings.py
@@ -72,7 +72,7 @@ class BackendSettings(BaseSettings):
         description="Database of users.",
         validation_alias="MEX_BACKEND_API_USER_DATABASE",
     )
-    redis_url: str | None = Field(
+    redis_url: SecretStr | None = Field(
         None,
         description="Fully qualified URL of a redis cache server.",
         validation_alias="MEX_BACKEND_REDIS_URL",