+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - add endpoints to get a single merged or single extracted item by id
+- add generic id filter to extracted, merge and preview endpoints
 
 ### Changes
 
@@ -17,11 +18,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
+- deprecate hadPrimarySource for extracted, merge and preview endpoints
+
 ### Removed
 
 ### Fixed
 
 ### Security
+
+## [0.40.0] - 2025-06-24
+
+### Added
+
+- BREAKING: add uniqueness constraint on provenance fields for extracted items
+
+### Changes
+
+- explicitly specify the location (Body, Path, Query) for all endpoint parameters
+- make redis_url a secret string
 
 ## [0.39.1] - 2025-06-19